+++ spec/grape/validations/instance_behaivour_spec.rb
@@ -6,7 +6,7 @@ describe 'Validator with instance variables' do
   let(:validator_type) do
     Class.new(Grape::Validations::Base) do
       def validate_param!(_attr_name, _params)
-        if @instance_variable
+        if instance_variable_defined?(:@instance_variable) && @instance_variable
           raise Grape::Exceptions::Validation.new(params: ['params'],
                                                   message: 'This should never happen')
         end